Send shard names from authserver to clients.
When sending the shard list from the authserver to
the client, also send along the names of the shards.
Unpack this on the client side and display them
correctly instead of only allowing the hardcoded
shard names.

The name is limited to 25 characters in authserver.
The client can't display more than that anyway.

This increases the authserver protocol version to
190516. This means that authservers using this protocol
version will not accept clients with an older version.

Patch outline courtesy of @BubbleWrap (I think)!