Fix #119; incorrect free of text buffer caused crash
When rendering the auction UI, the text description for
enhancement recipes is pulled to be displayed. However,
the code then tries to free the string in the
uiEnhancement which contains the description, but that
string is a StuffBuff and not an EStr, so the client
crashes.

Don't bother freeing it, as freeEnhancement will free
that buffer anyway.