Higher quality dynamically generated window icons
The text offset could be better, but it's a clear improvement over the previous solution in my opinion.
The default font (that attempts to get loaded) is Segoe UI since that appears to be Windows 10's default.
wchar_t was used so UTF-16 characters could be utilized in application icons. Hello emoji!

I made an effort to document the functions I touched, to hopefully set a precedent.