Copy some more project executables to bin directory
ChatAdmin: I've never used it, nor seen a use for it, but it can connect to and moderate ChatServer.
ShardMonitor: This is probably better off being a script. Or just documented a documented ServerMonitor flag somewhere... it launches ServerMonitor.exe -shardmonitor and quits.
TestClient: Useful for testing stability and various server functionality in an automated way.
TestClientLauncher: Control interface for TestClients.
DBQuery: Used by account portals to read real-time data.