Adds per-server port specification and allows clients to request specific server groups
* Implements new protocol OUROBOROS_PROTOCOL_VERSION_2 (191021)
This allows you to specify -servergroup X on the client command line up to 10 times, and instead of providing the normal list of servers, authserver will send the servers in the requested groups

Intended for use cases where a live and QA server are on the same authserver - assign live server to group 1, QA server to group 2, then add -servergroup 2 to the launch options for the second manifest, and users using the live manifest will no longer see the QA server and vice versa
Would only take minor changes if a server wants to remove the command line option and just hard-code a QA client to request a specific group instead

* Adds a column to the server table to hold the dbserver port.
If left at 0, uses the value specified in the config file.
The client already uses whatever port authserver sends it.
The port from the db is used regardless of protocol version, since client already expects to receive a port, we are just changing which one we send (sometimes).

Intended for use cases where two or more dbservers are running on the same IP. Requires the servers to be reconfigured to remove port conflicts, but that is out of scope for authserver and down to the server operator.