Added a config option to tell authserver to expect client server group requests, instead of checking protocol version directly. This is set automatically if the new protocol is in use, so new protocols can do the same if they want to keep this functionality.
Added a matching function in client that checks the protocol to decide if it should ask for specific groups or not.

Added a config option to tell AuthServer if it should try to read ports from the DB or not, to ease the upgrade path. Defaults to no, which is the existing behaviour.