Revisions per Clowd comments
Revised static numbers to defined values in timing.h, and included the file. Revised how calculations were done, because it wasn't working correctly before.

The game WILL NOT update the time played if you sit and type "/played" constantly. The values are not updated in the database until a mapchange event or log off/log in.