Added Aggro/Target cap to servers.cfg
New parameters "AggroCap <int>" and "TargetCapMode <int>" added to servers.cfg to make the aggro cap configurable and the target cap configurable.
AggroCap is looking for an integer greater than 17 (current cap).
TargetCapMode is looking for an integer of 1, 2, or 3.
1 - Default Target Caps: Reads the power definitions and maintains those maximums
2 - Scales the default target caps against the new aggro limit (Power Max targets/17)*aggrocap
3 - Removes the aggro cap (more accurately sets it to the aggrocap
Any other integer sets it to the default in powers as a catchall. Default value is 1.