Fix for major console error reports and incorrect gendered titles
A fix for an issue created by the original fix for gendered titles crashing the mapserver which, strangely, seems to have resulted in client crashes.   Client needs to use original handler, while the mapserver needs to use the updated handler.   Code should compile with the appropriate handler for the different applications.