Fix for /fx_special persistent effect zoning crash
Using /fx_special an admin can add any additional fx to a character, but previously if it was a continuing effect, the controlling client would crash when the character zoned - including when that character tried to log in again and every time afterward until the fx was manually cleared from the database.   The crash stems from a previous hack (cryptic's, not ours!) in the code that assumed all fx were associated with a costume piece.   A quick NULL check prevents that assumption from crashing the client, and allows continuous fx to be applied without an issue.