The return of HandlerE for gendered text
This is a shameful, shameful hack that rewinds us to the original fix for the mapserver crash related to gendered titles.   By way of this hack, we revert to using HandlerE on both client and server sides, which allows us to have the appropriate gendered title for the badge list, the visible title, and badge references in the chat box.