Added "split" as a name indicator for Worktables
I've expanded the worktable split view to search for the word "split" anywhere in the worktable name. By adding "split" it will change any worktable with that in the name to the split view with the level slider filter, and left side displays a drop down tabular menu, while the right shows the items in the selected tab. Previously this was limited to worktables with the following words:
ticket, incarnate, ham, vam

While these are re-usable, it makes the definitions a tad messy on the back end.