Feature: Allow server owners the ability to choose whether players of all classes can use petsay commands on commandable pets or just the Masterminds (default)          server.cfg : PetCommandsForAll 1 will turn this on, 0 will turn it off          Note- this also prevents people from using a hacked game client to bypass the restriction, as it turns out - the mastermind check was only carried out client side.          This would have been hard to exploit however, as I think the only commandable pets you get besides MM pets are the Lore pets- which is a level 50 thing.          This does not affect pets such as Singularity, Fire Imps etc
This feature should open it up for custom archetypes to have similar control over their pets as Masterminds do. Although that being said - while this is turned on - everyone will share the benefit.

Scenarios tested: PetCommandsForAll 1 + Defender with Ember Demon : Allows use of petsay
                  PetCommandsForAll 1 + Defender with Ember Demon + Fire Imps : Allows use of petsay on Ember Demon, but not on Fire Imps
                  PetCommandsForAll 1 + MM with Zombie : Allows use of petsay as normal
                  PetCommandsForAll 0 + Defender with Ember Demon + Fire Imps : Does not allow petsay on either
                  PetCommandsForAll 0 + MM with Zombie : Allows use of petsay as normal

Other scenarios that should probably be tested would include arena battles - as apparently your pets/gladiators were able to be petsay'd there.
This should be tested as the game client used to perform a validation of whether you were on an arena map on the same line as the mastermind check
The arena check has been moved to server-side also, but is done in a slightly different manner, re-using some functionality from elsewhere in the mapserver.
If anyone can help test the arena scenarios - much appreciated.